requirements.txt
Flask
Pillow
python-dotenv
mysql-connector-python
pydantic